refactor(management): deepen job action flow
